+++ lib/LoadCustom.dart
@@ -2,7 +2,6 @@ import 'package:aio_pc_controller/Custom.dart';
 import 'package:aio_pc_controller/LayoutSelect.dart';
 import 'package:flutter/material.dart';
 import 'DatabaseHelper.dart';
-import 'Dialogs.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 import 'CustomLayout.dart';
